refactor(sdk): split conn and stream modules
