+++ crates/sdk/src/lib.rs
@@ -4,20 +4,17 @@
 use std::{collections::HashSet, path::Path, sync::Arc};
 use tokio::sync::RwLock;
 
-pub mod client;
 pub mod command;
+pub mod conn;
 pub mod message;
 pub mod stream;
 
-#[cfg(unix)]
-pub use client::connect_uds;
-pub use client::{
-    ConnectionInfo, MemConnection, OutputChunk, Transport, connect_from, connect_mem, connect_tcp,
-    spawn_stream, stream_chunks,
-};
 pub use command::{COMMAND_HINT, COMMANDS, Command, collect_candidates, parse_command};
+#[cfg(unix)]
+pub use conn::connect_uds;
+pub use conn::{ConnectionInfo, MemConnection, Transport, connect_from, connect_mem, connect_tcp};
 pub use message::{Attachment, AttachmentKind, Message, attachment_summary};
-pub use stream::StreamAccumulator;
+pub use stream::{OutputChunk, StreamAccumulator, spawn_stream, stream_chunks};
 
 /// Shared set of sender IDs belonging to sibling Crabtalk bots.
 ///
